refactor: move search state to entriesSlice
When adding a new entry, the filter needs to be
re-sorted so that the new entry will be hidden if
it is not fit for the search criteria. Accessing
an external reducer seems to be a pain to do,
and if you need to access another reducer then
the general consensus seems to be that you need
to restructure your state.

So instead of having the search functionality in
its own slice, I decided to combine it with the
entries' slice since entries are directly affected
by the search terms.